Refactor AdventyPayButton component: move to new directory, update styles, and enhance tooltip functionality
